Remove everything code, keep Twig Templates (#33)
This PR removes all the machinery that was necessary in early Symfony 2.x versions to preview error pages during development. We've been able to get this functionality merged into Symfony 2.6 (see https://symfony.com/blog/new-in-symfony-2-6-error-page-previews).

So, it's time to let go and remove everything except the Twig templates that still provide useful building blocks. Not having the code in the first place makes it easier to maintain this bundle for newer Symfony versions.